Check find's output in Jenkins' translation scripts.
* modules/jenkins/files/slave_scripts/propose_translation_update.sh:
Instead of checking find's exit code check finds output to see if it
found any .po files. This is a more reliable way to check for the
presence of .po files.

* modules/jenkins/files/slave_scripts/upstream_translation_update.sh:
Instead of checking find's exit code check finds output to see if it
found any .po files. This is a more reliable way to check for the
presence of .po files.

Change-Id: I53f7cc47e3762e4529eaf72f457852e650413b16