add ep_fintest, needed by ep_headings to etherpad
Make it possible to optionally install ep_headings on an etherpad
environment. This makes it easy to enable this on some environments,
but not all. Default to 'false', but set 'true' for etherpad_dev.

on my environment I also had the installation of ep_fintest, because
plugins didn't seem to work until I did this. Testing that.

Change-Id: Ia1ee0fd7b51e88f639de5350f124f1a95405cabf