Remove absent directories forcibly.
* manifests/site.pp: Setting the ~jenkins/.config/glance directory
recursively absent on unit test slaves in 35ccdfc was an
insufficient measure to remove it and its contents. According to
Puppet documentation force must also be used to remove the file
resource if it is a directory type, so this change adds that
parameter.

Change-Id: I834c229b83d6ffd8c0c1a3f4912699a775e01825
Reviewed-on: https://review.openstack.org/26482
Reviewed-by: James E. Blair <corvus@inaugust.com>
Approved: Clark Boylan <clark.boylan@gmail.com>
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Tested-by: Jenkins