Clone requirements into a tempdir.
Git doesn't remove git repos with git clean, so the workspace
checkout of the requirements repo was stale.  Clone into a
tempdir instead.

Also, fix a traceback in the case that a new requirement is added.
The loop wasn't short-circuiting as it should in that case.

Print the git sha from the requirements repo to aid in debugging.

Reorder imports.

Change-Id: Id5a8e748a1fb353f2628a9c9fa291d6825d1046d