Add zmq-event-publisher Jenkins plugin to Gerrit.
Create a new project under openstack-infra for the zmq-event-publisher
Jenkins plugin. We are actually beginning to use this plugin for
logstash so it should follow our normal processes for code review and
merging.

The gearman-plugin project uses the same sort of build and package steps
so these Jenkins jobs have been converted to a set of templates in a job
group that both plugins can use.

Change-Id: I2bab79506119b1c04a8e9d48bc66f1a609426324