Puppet dashboard mysqld innodb_file_per_table.
In 232747f we attempted to implement it this way initially, but
subsequent patchsets went off into the weeds instead trying to use
mysql::server::config which, as it turns out, was not introduced
until a newer version of the puppetlabs mysql module than we use.
Rather than go through a potentially disruptive infrastructure-wide
upgrade right this moment, this patch gets the job done in a more
traditional fashion.

Change-Id: I6967388072c7c8b3e170048e23618ea95bccfecc