Add post-mirror-python2X jobs to the post queue.
* modules/openstack_project/files/jenkins_job_builder/config/mirror.yaml:
Add new post-mirror-python26 and post-mirror-python27 jobs to the mirror
jobs config. These jobs will run in the post queue and ensure that the
logs for these jobs end up in the correct location. Previously we had
run the periodic jobs in the post queue which copied the logs to the
wrong place.

* modules/openstack_project/files/zuul/layout.yaml: Run the
post-mirror-python2X jobs in the post queue instead of the
periodic-mirror-python2X jobs.

Change-Id: I26f38d937a55dcd61ac17b5627ffb3eb1e22d63c