Add gate job configuration for Chef cookbooks
This commit adds the Jenkins job builders and gate
configuration for the Chef cookbooks. Only the
cookbook-openstack-common repo is changed from a
NOOP gate to the lint + unit gate at this time, as
that cookbook is the first that will be going into
the Stackforge repo.

Change-Id: I999a990f15829730495ab0891e4d38f1a5795d60
Reviewed-on: https://review.openstack.org/28640
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Reviewed-by: Khai Do <zaro0508@gmail.com>
Reviewed-by: Jeremy Stanley <fungi@yuggoth.org>
Approved: Jeremy Stanley <fungi@yuggoth.org>
Tested-by: Jenkins