Rework jenkins puppet module packages
Feel free not to accept this.  I added ruby1.9.1 support to debian
jenkins slaves, and found params was doing OS detection, yet the
slaves manifest was doing more package OS detection.  We are already
maintaining two sets of lists.  Rather than abstracting away the
package name into variables, only do this when necessary.  I am on
the fence as to which way is better, however, this way feels slightly
better.

Change-Id: I29715f94a97898f5275f45eba322d0a6d255cdd0