Remove whoopsie package from all servers.
We've seen the whoopsie process eating many gigs of virtual memory
on servers, and it has a history of memory leaks. It's also more of
a desktop-oriented application geared for reporting backtraces to
Canonical when programs crash, so we don't need it on our servers
anyway.

Change-Id: I83a16b557d2601acdfcbe4c342bb05d325df77c9
Reviewed-on: https://review.openstack.org/29117
Reviewed-by: James E. Blair <corvus@inaugust.com>
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Approved: Clark Boylan <clark.boylan@gmail.com>
Tested-by: Jenkins