Clarify StackForge document audience and workflow.
* doc/source/stackforge.rst: Adjust the project onboarding sequence
to suggest requesting an initial group member last, since this seems
to be the simpler and less error-prone order. Also add a brief
subsection to the introduction describing the intended audience and
linking to prerequisite documentation for those who need some
additional education on OpenStack code review process and
convention.

Change-Id: I4a7cc6e56505a462c5596abe587c560bfa722f00