Set an upper bound to log pusher output queue.
Logstash's queues are small. During periods of high load this can cause
the log pusher to cache all of the logs it is pulling in. Set an upper
bound to its output queue size to force it to block when Logstash cannot
keep up. This should prevent OOMing and won't impact log throughput as
Logstash is blocking anyways.

Change-Id: I9ca31e6dbe454e9c4878fd7ba35bc33bc9df7d83