Force releasestatus to use IPv4 for SSH to Gerrit.
At random, some hosts within Rackspace refuse to send IPv6 packets
with DSCP marker 0x10 (lowdelay). This is the default behavior of
recent OpenSSH servers and clients, and results in SSH sessions
hanging on these hosts. When static.openstack.org was rebuilt a few
months ago, we did not immediately realize it suffered from this
problem as Rackspace thought they'd fixed the issue previously. We
have an open trouble ticket with them for the past several months to
get it addressed properly, but have no resolution yet. In the
meantime, we have this lovely workaround...

* modules/releasestatus/files/ssh_config: New SSH config file forces
IPv4 when connecting to review.openstack.org from the releasestatus
account.

* modules/releasestatus/manifests/init.pp: Installs the new SSH
config file.

Change-Id: I2968fc00750e3a5a23c08a3007318fd2e336b251