Increase python unittest timeout to 40 minutes.
Similar to cd02e47, the python stable bitrot unittests (particularly
python26) have been getting close to the test timeout and
occasionally go over. Increase the test timeout by 10 minutes for a
timeout of 40 minutes. These jobs run daily from a timed trigger so
should not increase the runtime for changes in the qate.

Change-Id: Ide580777c0fb062ebce3aa066d5647062009a5bd