Don't publish non-master docs to master location.
The current run-docs script only moves the publishing location on
tags or stable branches.  This makes it so that master docs stay
where they are, stable branches and tags have the current behavior,
and anything else is placed in the branch name.  This should affect
milestone-proposed (which will be published to .../milestone-proposed/)
and feature branches (.../feature/foo).  Currently, these two cases
are in a race with master -- last change wins.

Change-Id: If1e09d31a7001255746ba1a91b641b95f42fcfbb