pass version info on every jenkins plugin build
This commit is to fix bug 1182154.  This commit sets up a version
string as an environement variable so we can pass it to maven builds.
This makes the build workflow similar to how we build python packages.

The version string is retrieved from information in git.

This setup expects a variable called '$project-version' in
the root 'version' element (i.e. ${project-version}</version>) of
the project's pom.xml file.

For general (throw away CI) builds we do the following:
  1. generate a package 'myplugin.hpi' with version '1.3.0.4.a0bc21f'
     in the MANIFEST.MF file.  The '4' is the number of commits since
     last tag.
  2. publish 'myplugin.hpi' to tarballs.o.o

For release builds we do the following:
  1. generate a package 'myplugin.hpi' with version '1.3.1'
     in the MANIFEST.MF file.
  2. publish 'myplugin.hpi' to repo.jenkins-ci.org
  3. publish 'myplugin-1.3.1.hpi' to tarballs.o.o

Change-Id: I3e75b677889b50b515233637a32a424bc5b7a41b