Initial configuration for Barbican
Changes needed for the StackForge Puppet process to create Barbican project.

Change-Id: I78c1ed0f0033f61b3a88dce6eda100933d13f910
Implements: blueprint create-in-stackforge:

Attempt to fix layout.yaml configuration issue

Based on the Jenkins error report, I think it is an error to specify
a 'publish' 'gate-noop' setting, so I am removing it and adding a patch.

Change-Id: I1f63ce5e96b9a007a07b8bcc5534b20bbb7d4aa1
Implements: blueprint create-in-stackforge
Fixes: bug #1188660