Add job to test project installation.
Test that each of the projects can be installed from scratch without
pbr being previously installed into a plain raw non-distribute-based
virtualenv. This is the closest we can come to testing that pip
installing from the internet is going to work.

First we have to create a local fake pypi mirror that contains the
current version of pbr and its depends, because the initial setup.py
install is going to be the source of installing pbr, and we don't have
the candidate version of pbr in our mirror yet. (duh)

We ALSO need to test that we can do these things with the current
released version of pbr.

Change-Id: I7b9a4299d1d42a7e4bd3c7e1895a02d2493f158a