Fix navigation order with HOWTOs
Move the toctree from the HOWTOs sidebar on the front
page to a hidden toctree after the main toctree.
Create manual links to the HOWTOs docs in the sidebar.

This fixes the fact that the "next" link in the docs
navigated to one of the sidebars first, instead of the
introductory "Infrastructure Project" doc, while keeping
those sidebars in the toctree navigation so if someone
reads the docs straight through they get every page.

Change-Id: I5e1c23b7d3ce324bbd9e333cf6e898b33601982e
Signed-off-by: Doug Hellmann <doug.hellmann@dreamhost.com>