add help infrastructure and initial stub to logs
in order to make it simpler for people to understand the results
of a devstack/tempest run, provide in line explanation for the
runs via an apache directive. This will let us sprinkle appropriate
help for different tests and directories on the log server.

DirectoryMatch fails to work with ReadmeName, but Directory with
creative globbing works (thanks to experiments by fungi).

Change-Id: I52e648dac6f13babd01ea245414ef54c243cdd36