Fix setting of properties
running the bash script with a properties build wrapper doesn't work
because the wrapper runs before a scm checkout since the
script queries git it needs to be run after checkout.  This
commit uses the shell builder to run the script then the
evninject builder to read the generated properties file
to make it available to the rest of the build.

Fixed the script to always return 0 exit code on git commands.
Also changed $() to back ticks for consistency.

Change-Id: I1f31076762fb1d33d7e889fd2de4cf28d64d922d