workaround for jenkins bug https://issues.jenkins-ci.org/browse/JENKINS-14193
The jenkins maven builder cannot reference environment variables if it is set
by the EnvInject plugin at the prebuilder step.  However the maven builder CAN
reference variables when loaded from the build environment setup (wrapper) step.
The only problem is that gerrit-git-prep runs in prebuilder step which occurs after
a build environment step.  We want the script to run after gerrit-git-prep so
we are calling the gerrit-git-prep in script so it will run in correct order.

Change-Id: Iedd7561b4ef8e19b3713b63bb76b9b754ce116e9