Don't try to compress patterns ignored by gzip.
* modules/openstack_project/files/log_archive_maintenance.sh:
According to the gzip manpage, there are a variety of filename
endings it will consider to indicate a file is already compressed
and will then refuse to compress. Some tempfiles get
randomly-generated names which match these endings, causing the
compression cron job to emit errors. Instead we make the script skip
over them.

Change-Id: Ib78a19362eaa54c43f44a9e6bd0c265dddee92be