Add support for bare puppet modules to puppet-lint
puppet-lint expects the parent folder where the test is executed
to be named after (or ends with) the module name.

If the folder isn't named correctly, puppet-lint raises an error:

    "XXX not in autoload module layout on line YY".

This change adds support for bare puppet module projects by moving
all files in a new folder named after the project. The project name
is expected to end with the actual module name for it to work.

If a "modules" folder is found, nothing is done. We suppose
the project is using a multi-modules layout and puppet-lint runs fine
against this kind of layout.

Change-Id: Id9e8a7edb8d64046e6e57aa4c161a9947edc25d9