Enable the cells testing job
This enables the cells testing job.  It's non-voting for now.  It's also
only running during check and not gate, so it's not unnecessarily
wasting resources while it's still non-voting.  We can make it voting
once we're confident that things work and are stable.

Change-Id: Iaa8a0e7d3d36e6703e96497247208d60453dcfaf
Reviewed-on: https://review.openstack.org/35544
Reviewed-by: James E. Blair <corvus@inaugust.com>
Reviewed-by: Jeremy Stanley <fungi@yuggoth.org>
Approved: Clark Boylan <clark.boylan@gmail.com>
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Tested-by: Jenkins