Use Modulefile to get module name for puppet-lint
Previous method used '{project}' to determine the module name.
It wrongfully assumed '{project}' would get substituted for
the project name by Jenkins job builder.

This change uses an alternative method to determine the module name
by reading the 'name' field from the Modulefile.

If the Modulefile is not found or the name cannot be determined,
the puppet-lint builder falls back to its default behavior
which consists of running puppet-lint in the root of the workspace.

Change-Id: I6cd8fce6052c94fc0d3c3f75ee0a62c366e430e6