Import sqlalchemy-migrate
Upstream decided that they did not have time or interest in
sqlalchemy-migrate anymore. Since we still have a dependency on it, and
since we also have an upcoming need to upgrade to sqlalchemy 0.8, we
agreed to take over its care and feeding.

The overall project plan is to transition to alembic, so at some point,
we will no longer care about this tree, but that time is best-case over
a year away.

Change-Id: Ib133190a01fcb12122a58c544dfacfeb3e2c107b