add utility to htmlify screen logs for devstack runs
this adds a tool to jenkins to htmlify our screen logs, adding
spans with classes that allow the output to be colored, and adding
anchor tags that make it possible to hot link into a log based on
a timestamp. It should work for all logs that use common olso
logging for log formats.

Change-Id: I12a06c4b55be05fb8060b76f5aea9ebc801c0537