Set up basic Asterisk configuration
This commit sets up the basic configuration for Asterisk.  It will allow
Asterisk to run, but it won't do anything useful yet.

The asterisk class comes with a large set of default configuration files
with sane default values.  Most of these files are not actually needed
for our immediate use case, but we might as well have a complete set for
making the asterisk class as generic as we can.

The node must specify its own modules.conf and asterisk.conf files.
Everything else is handled using a conf.d mechanism.  For each file we
want to customize, we will drop files into the conf.d directories as
applicable.

Change-Id: I7975082ff5351db4dc6e3c8cf9dd2d90675e3108