Enable basic conferencing support
Set up basic conferencing support.  Right now I have reserved 6000-6999
as conference rooms (not that we actually need that many, but whatever).

At the time of writing, the sound packages need to get updated because
they are missing the confbridge sound prompts, which are needed for this
to work.  I'm working with the package maintainer to get that done now.

Change-Id: I9acddf4ffedc7f499740184778b8bd67e5b38a4f