Work around catch-22 in jeepyb with broken pbr
* modules/jeepyb/manifests/init.pp: Jeepyb depends on pbr and
installing or upgrading from the local source tree relies on the
already-installed version of pbr before it can pull in a new version
of pbr. That leaves the jeepyb puppet module open to breakage if the
installed version of pbr happens to be particularly fouled up. Work
around the potential issue by manually upgrading pbr first.

Change-Id: I7154aebead23f9dd22a89f1cbd3da91a2832da31