Conditionally install Python 3 development libs.
* modules/jenkins/manifests/params.pp
* modules/jenkins/manifests/slave.pp: The conditional check introduced
in the Jenkins slave manifest in commit 3c7af7d was only a half-fix
since the python3 pip provider lacked a similar check, so depending
on the order in which it's puppeted the conflict could still arise.
Since this package is not needed on non-py3k slaves, we can just
remove it from the slave manifest and expect slaves which do need it
to get it from the pip puppet module instead.

Change-Id: If22aaa1319b61aa58306036a03843fa0cc79b34c