Conditionally install Python 3 develoment libs.
* modules/pip/manifests/python3.pp: The conditional check introduced
in the Jenkins slave manifest in commit 3c7af7d was only a half-fix,
so depending on the order in which it's puppeted the conflict could
still arise. This similarly checks within the pip provider to make
sure Python 3 development libs are only installed via distribution
packaging if they're not already defined elsewhere.

Change-Id: If22aaa1319b61aa58306036a03843fa0cc79b34c