Create new repo to host legacy heat-cfn client.
Our cfn compatible clients (heat-cfn, heat-boto, heat-watch)
need to be moved out of the main heat server repo.

Activity on this repository is expected to be low until cfn compatibility
becomes a priority. However this client is still important enough to be
in its own repo and have its own release cycle.

Change-Id: I5a1f8cb51f20733fa9b91ebbd60fe3066e3313f8