Record the files present in the mirror.
TripleO wants to be able to easily reproduce the OpenStack CI PyPI
mirror. We can do this efficiently in an unprivileged fashion if the
contents of the mirror are provided as a concordance. In this change I
list all the files that a particular mirror run will produce into a
unique file for that particular run. We can merge those easily on the
client side of TripleO's disk-image-builder.

Change-Id: Ib6c7a8864079fffcacdc2472571b8ee0c5a15e5d