Increase d-g launch timeout to 20 min
The script allows a node 15 minutes to boot.  Additionally, a bit
of time is needed afterwords to add it to Jenkins.  Increase the
timeout of the launch job from 10 to 20 minutes so that entire
process is more likely to occur within one run of the job.

The job itself is still scheduled to run every 10 minutes, so if
this proceeds quickly, we will still start a launch job every 10
minutes.  If things are slow, we may only run it every 20 minutes
in the worst case.  The job is not allowed to have concurrent runs
so this is safe, and the n+1 job will simply be queued.

Change-Id: I879c4d309116f1b0517d878b7365ebc29b27f547