float the progress bars and statuses to the right
this floats the progress bars and statuses to the right, which
provides a nicer visual block for all the status elements. The
progress bars need a margin hack to get them to visually align with
the rest of the line.

This also moves the progress bar width into the css instead of
inline, and to 5em which is approx what SUCCESS width is in chrome.
(Should be about the same in other browsers, but font rendering
is different between webkit / firefox).

Errant white space deleted, my editor does that on all changes,
sorry if that noises things up.

Change-Id: Ibf4288049e3b24949d17a6b39f2cf1b7bd7a52dc