make oslo.config gating on devstack/tempest runs
oslo.config from git is going to be used in our devstack/tempest
runs, so ensure that we're gating on it in a matrix of configurations
that cover all they ways that an oslo.config compatibility problem
could cause a project fail, which could wedge the gate.

This is WIP until we get grenade sorted to ensure it works there.

Change-Id: I2ee6acd0073b284ab6ba61fc211562ccf55a5320