Pass testr and subunit2html.py a PYTHON envvar.
* modules/jenkins/files/slave_scripts/run-tox.sh: Some subprocesses
of testr get invoked from the calling environment's configured
interpreter regardless of what interpreter was used to invoke the
parent process. Put that in a PYTHON envvar instead of on the
command line.

* modules/jenkins/files/slave_scripts/subunit2html.py: Fix this
script to invoke the interpreter specified by the calling
environment instead of using an explicit executable path in its
shebang.

Change-Id: Id27e4a280e88b2ce386af6e9f248b1c4e004583a