Update zuul to be compatible with the latest apache module
Previously, the zuul module required an extremely old version
of the apache module (0.4.0). The usage of that module has
changed quite a bit since then, and using the latest version
requires code changes.

DO NOT MERGE. This is just an example of how to do it, but since
other modules rely on the older version, this must be part of a
bigger refactor.

Change-Id: I71f28740c88653597cc058889b8b043f0dc43c1e