Add shell script syntax checking job
ksh -n, weirdly, can be run on bash scripts to find syntax errors and
also warnings. Add a script to use this to run syntax checking on shell
scripts, then add it as a job to the config repo and clean up the
warnings it produces. I think if this goes well, considering it for
devstack might me nice too.

Change-Id: I9f0a610ea693b86a7b53af1f5cbc9e425f70dbba