Prepare to test git-review
* .../files/jenkins_job_builder/config/projects.yaml: Switch the
python-jobs group in for the previous pep8 and tarball jobs, since
we want to start running pythonXX tests too. Also add the python33
job because it's not part of that group.

* .../files/zuul/layout.yaml: Run voting gate-git-review-pythonXX
jobs. These are effectively no-op tests for the moment, but will be
shortly followed by actual integration tests already under review.

Change-Id: I8cd516eef55e25d200609f6de3f06d322df383c7