Add support for per-distro wheel mirrors
Because the wheel format does not do fine-grained enough selection on
things like the high likelihood that Ubuntu precise and Ubuntu raring
might have different underlying C libraries for python 2.7, we need to
put our built wheels into a structure that allows them to be used
safely.

lsb_release -r -i is used because Centos lists lsb_release -c as "Final"
which is not particularly helpful. This will wind up with identifying
strings like Ubuntu-12.04 and Centos-6.4 - which should be good enough
for what we need. The same logic is used in run-mirror to produce the
subdir.

Change-Id: I84c0ef57cf2e5e8c2cbfe2b18512b29228b0838f