Don't compress very, very recent log files
* modules/openstack_project/files/log_archive_maintenance.sh: On
occasion the compressor will be engaged on a file which is still
being written to the log archive, resulting in a compressed but
truncated copy along with the original it lacked the write lock to
unlink. Then on a subsequent pass, this file will be skipped because
the destination filename already exists. Instead, don't compress a
file unless its last modified time is more than 10 minutes in the
past.

Change-Id: Ife874784babd41aa45b4183f4f0a5a5730cf5a23