Switch etherpad_lite backups to mysql_backup.
* .../openstack_project/manifests/etherpad.pp:
Use the new mysql_backup module to backup the etherpad DB. Note that
this change is not sufficient to completely clean out the old logrotate
config and cron for the old etherpad DB backups. This should be merged
in a coordinated manner where the other logrotate configs and cron are
removed by hand.

Change-Id: Ide60ce13a454f72865d697ee736beada76850f32