Add job to propose tag merges
With pbr tag-based versioning, not having the latest release tag in the
master branch makes the output a little nasty. Add a job that makes a
merge commit containing the tag on a release from milestone-proposed
branch. The merge strategy it uses is 'ours' because we're not
attempting to use this to keep code lines up to date - the expectation
is that content added in milestone-proposed should already be in master.
The main thing we want to do is re-connect the revision and tag history
with master, so that things like 'git tag --contains' work.

Change-Id: I06c06bd7542ae493e74fdcaa3c4419f84843cd03