Fix intermittent jenkins plugin build failure
* modules/openstack_project/files/jenkins_job_builder/config/jenkins-plugin-jobs.yaml
Changed from maven job to freestyle job. The freestyle job builds with maven
so that makes an assumption that maven must be pre-installed on the build machine.
Changing to freestyle job removes the need to have a work around for JENKINS-14193 bug

* modules/jenkins/manifests/params.pp
Make sure maven is installed on build machine.

* modules/jenkins/files/slave_scripts/maven-properties.sh
remove workaround for jenkins bug

Fixes: bug #1212470
Change-Id: Ia78ea699c70efc08f6dd93bde321faece99900f1