Add cookiecutter templates repo
Cookiecutter is a project that creates skeleton projects from a set of
templates. Since there are at least a few different files to set up for
an openstack-compliant project, let's maintain a set of templates.

After this has landed, we should add support to manage_projects to use
cookiecutter to create the initial empty repo.

Change-Id: Iad67d289df1d37c1e8172e18b4903ca8ab278fa5