Proxy git-daemon with haproxy.
git-daemon is really good and DoSing the git server. Put it behind
haproxy so that we can define the max number of connections and anything
beyond that will queue. This should give us a good balance between
faster git protocol and not DoSing the server.

* install_modules.sh: Add puppetlabs-haproxy forge module.

* modules/cgit/files/git.xinetd: Relax ACLs so that all connections
coming from a single source do not cause problems.

* modules/cgit/manifests/init.pp: Add haproxy config, depend on xinetd
service.

* modules/openstack_project/manifests/git.pp: Enable haproxy.

Change-Id: Icefc5923cff9a7c6ce62c1923ec2ea87ebc6474a