Disable tempest in the cells job
As we approach the feature freeze, I know jenkins will be taxed for
resources.  Since tempest isn't passing for the cells job right now,
let's just disable it for the next couple of weeks at least so that this
job finishes quicker and helps free up a little bit more resources.

It's still quite useful to have the devstack exercises run, because at
least we're getting a basic instance build done with cells.

While we're at it, disable this job completely on most projects to
further free up resources.  This leaves it running for nova, novaclient,
devstack, and devstack-gate.  It seems *incredibly* unlikely that
changes elsewhere could break cells specific code, so let's just not run
it elsewhere for now.

Change-Id: Ice513ca9d424b0e0082121dba37a5165d71760cd