Disable tempest in the cells job temporarily
As we approach the feature freeze, I know jenkins will be taxed for
resources.  Since tempest isn't passing for the cells job right now,
let's just disable it for the next couple of week so that this job
finishes quicker and helps free up a little bit more resources.

It's still quite useful to have the devstack exercises run, because at
least we're getting a basic instance build done with cells.

Change-Id: Ice513ca9d424b0e0082121dba37a5165d71760cd