Document setting the correct hostname.
Puppetmaster bringup will often have the wrong hostname in a cloud
(e.g. a uuid or the shortname given to nova). But if it's wrong the
puppetmaster CA created by the puppetmaster module will be broken.

Additional nodes can have this handled by automation, but the
puppetmaster is special - this is part of the bootstrap process.

Change-Id: I8e0e2ea4a5d27fb248b88d6be57abb171d81f13f