Give Jenkins the ability to trigger puppet runs
* modules/openstack_project/files/salt-trigger.sudoers: Allow the
jenkins user to send messages to the salt master.

* modules/openstack_project/manifests/salt_trigger_slave.pp: Add the
sudoers inclusion above to the salt-trigger slave.

* modules/salt/manifests/master.pp: Change ownership on the existing
salt configuration directory on the master to belong to the salt
user, and add the file structure for the new reactor components.

* modules/salt/templates/master.erb: Add reactor configuration to
run tests.sls when receiving a trigger named jenkins.

* modules/salt/templates/tests.reactor.erb: Define a puppet command
which will be run on all minions.

Change-Id: I346bb28e5b4d53618855a28f616f7c5ed0e60dc7