Remove closed bugs from recheckwatch
Once a bug is closed there is no reason to show it under recheckwatch.
If the bug was closed incorrectly then someone should re-open the
bug, and then use recheck. The benefit is that most of the time once a
bug is closed it is not relevant to recheck. Reducing the number of bugs
listed in recheck making it easier for a patch author to
find the relevant one in the list.
Right now there are 48 closed patches listed on recheck out of 132 bugs,
so this will reduce he number of recheck bugs by about 25%.

Change-Id: I48d4056f70a2be37af51c01547d1274566c22734