Visualize Zuul's NNFI scheduler
Adds a subway map to the status page that reflects the current
proposed merge state according to Zuul's NNFI scheduler.

Makes the screen wider (1024 instead of 950 px) to accomodate
the additional width of the subway map.

Widen the graphs as well to match (also, evenly space them).

This isn't nearly as generalized and elegant as I'd like, but
it should do as a first cut in order to see what's going on.

Change-Id: I11689c24c5d4aa58bb2dac3c595068e24af7e1d5