Test nova changes against a real database
The gate-real-db-upgrade_nova_mysql is implemented by turbo-hipster
https://github.com/rcbau/turbo-hipster which runs migrations and
checks for errors against a snapshot from a real database.

Currently there I have configured one worker up to test the gate
which is why this is only initially being added to the silent
pipeline. Reports are sent to me so I can see the result being
sent back to zuul.

Change-Id: I361ec15e7c3bc703035c1fa87b83c7bf10641017